add section lines
